filter primers complete
I have gotten the filter primers function completed. I have also dome some house cleaning so there might be a few more files than normal. One notible chance is that I have made the functino of get inside the primer always return a string. So it will need to be interpreted by the python module if you want to use it as an integer, etc. I will look into fixing this further sometime either this week or the next.