Added Missmatching
I added the function without any testing, It should work but besides some thrown errors, I need to go back through and add checks before I try to get the mismatch. There is also not a great way of testing this at the moment, I am hoping that the project testing would be completed in a python environment for simplicity.