Generate_maching_primers innit
Started working on it, there are errors so this function does not work at the momnet. I an running into issues with some of the variable names and types and trying to keep them consistnent from the reference I have. Also this function might not be possible because it is depedent on using the evaluation function which uses the primer3 library. I am planning on reaching out to the developer of the primer3 library to see if I could add the functionaility to our library...