Build seperation & dataframe to primer
I have changed the files needed to build the library for windows and linux. I also got a function together that converts a pandas dataframe into our primer, I am considering adding an output function as well to let you output as a dataframe. This also taught me that I can simply run a python command inside the c++ project if need be, which will be helpful for the event that we need to use primer 3, etc.